requirements_pdf.txt
# PDF Extraction Dependencies
# For text extraction
pdfminer.six>=20221105

# For table extraction - choose one of these options:

# Option 1: Full camelot with OpenCV (recommended)
camelot-py[cv]>=0.10.1

# Option 2: If OpenCV installation fails, use base version
# camelot-py[base]>=0.10.1

# Alternative/additional table extraction
tabula-py>=2.5.1

# Additional dependencies that might be needed
pandas>=1.5.0 